Bare bones of internationalisation effort
This is a test for internationalisation of Verona. It's not complete nor
tested and isn't integrated with the rest of the project (neither CMake
nor external use).

The basic idea is to have multiple languages, a default base language, a
way to read the files, a way to get messages by key. There is no
formatting yet as checking types gets complicated.